requirements.txt
fastapi
uvicorn[standard]
pyyaml
psutil
python-multipart
bcrypt
websocket-client
PyJWT
pytest
